requirements.txt
mcp-python>=0.1.0 